Task 3/requirements.txt
# Python Basics
wheel
pillow
python-dotenv
PyYAML
ipython
requests

# Django
django==3.1
djangorestframework
